Organize baseline scripts by pipeline
