+++ baselines/heat_eq/pinn_mlp.py
